Initial commit.
Python script also creates a file of the censored source text as well as print to stdout but this can be removed if required.
Additional memory can be saved by only considering one line of the text file at a time, but this is only feasible if it's reasonable to load in all censor words to RAM at once. This implementation allows the number of censor words considered at once to be set, meaning lines have to be revisited in the censoring process for each new set of censor words being considered.